+++ src/text/fmt/buf.rs
@@ -3,14 +3,17 @@
 //! Non-allocating formatting backed by a buffer.
 //
 
-use crate::{fmt_write, iif, FmtArguments, FmtError, FmtResult, TextWrite, _core::cmp::min};
+use crate::{iif, FmtError, FmtResult, FmtWrite, _core::cmp::min};
 #[allow(unused_imports, reason = "±unsafe")]
 use ::core::str::{from_utf8, from_utf8_unchecked};
 
 /// Returns a formatted [`str`] slice backed by a buffer, non-allocating.
 ///
-/// It calls the [`format_buf_args`][format_buf_args()] function with the
-/// [`format_args`] macro.
+/// Underneath it calls [`Fmt::format_buf`][crate::Fmt::format_buf] and [`format_args!`].
+///
+/// See also the allocating [`format!`] macro.
+///
+/// [`format!`]: crate::format
 ///
 /// # Example
 /// ```
@@ -19,46 +22,27 @@ use ::core::str::{from_utf8, from_utf8_unchecked};
 /// let s = format_buf![&mut buf, "Test: {} {}", "foo", 42];
 /// assert_eq!(Ok("Test: foo 42"), s);
 /// ```
-///
 /// # Features
 /// Makes use of the `unsafe_str` feature if enabled.
 #[macro_export]
 #[cfg_attr(cargo_primary_package, doc(hidden))]
 macro_rules! format_buf {
-    ($buf:expr, $($args:tt)*) => { $crate::format_buf_args($buf, format_args![$($args)*]) };
+    ($buf:expr, $($args:tt)*) => {
+        $crate::Fmt::format_buf($buf, $crate::format_args![$($args)*])
+    };
 }
 #[doc(inline)]
 pub use format_buf;
 
-/// Returns a formatted [`str`] slice backed by a buffer, non-allocating.
-///
-/// See also the slightly more convenient to use [`format_buf!`][crate::format_buf!] macro.
-///
-/// # Example
-/// ```
-/// # use devela::text::format_buf_args;
-/// let mut buf = [0u8; 64];
-/// let s = format_buf_args(&mut buf, format_args!["Test: {} {}", "foo", 42]);
-/// assert_eq!(Ok("Test: foo 42"), s);
-/// ```
-///
-/// # Features
-/// Makes use of the `unsafe_str` feature if enabled.
-pub fn format_buf_args<'a>(buf: &'a mut [u8], arg: FmtArguments) -> Result<&'a str, FmtError> {
-    let mut w = WriteTo::new(buf);
-    fmt_write(&mut w, arg)?;
-    w.as_str()
-}
-
 #[derive(Debug)]
-struct WriteTo<'a> {
+pub(super) struct WriteTo<'a> {
     buf: &'a mut [u8],
     len: usize,
 }
 #[rustfmt::skip]
 impl<'a> WriteTo<'a> {
-    fn new(buf: &'a mut [u8]) -> Self { WriteTo { buf, len: 0 } }
-    fn as_str(self) -> Result<&'a str, FmtError> {
+    pub(super) fn new(buf: &'a mut [u8]) -> Self { WriteTo { buf, len: 0 } }
+    pub(super) fn as_str(self) -> Result<&'a str, FmtError> {
         if self.len <= self.buf.len() {
             #[cfg(any(feature = "safe_text", not(feature = "unsafe_str")))]
             { from_utf8(&self.buf[..self.len]).map_err(|_| FmtError) }
@@ -70,7 +54,7 @@ impl<'a> WriteTo<'a> {
         }
     }
 }
-impl TextWrite for WriteTo<'_> {
+impl FmtWrite for WriteTo<'_> {
     fn write_str(&mut self, s: &str) -> FmtResult<()> {
         // IMPROVE?
         let rem = &mut self.buf[self.len..];